requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
google-generativeai==0.3.2
psycopg2-binary==2.9.9
langchain==0.0.352
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0 